Add files via upload
These codes utilise three different discretisation schemes for modelling the transport property in Computational Fluid Dynamics (CFD). The 3 schemes are, Central Difference Scheme (CDS), Upwind Difference Scheme (UDS). and Power Law Difference Scheme (PLDS). The schemes are compared to the analytical solution of the problem and the error is found for each scheme. The code is built so that the u, Gamma, and N values are tested and adjusted.